refactor: @Transactional 범위에서 외부 I/O 분리 - AudioService.saveTimeStamp()의 @Transactional 제거 - DB 쓰기 로직을 AudioPersistenceService로 분리 - 이에 따라 OCI 업로드 및 Whisper API 호출 중 DB 커넥션이 점유되는 문제 해결
-detached entity 문제를 방지하기 위해 Article 객체 대신 articleId를 전달하고 트랜잭션 내에서 재조회하도록 수정.
